pyproject.toml
[project]
name = "playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic[standard]>=1.18.3",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.128.4",
    "passlib[argon2]>=1.7.4",
    "ptpython>=3.0.32",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.1",
    "scalar-fastapi>=1.6.2",
    "sqlmodel>=0.0.32",
]
